30.5 print

+++ Project/Dragons.h
@@ -5,7 +5,7 @@ class Dragons :
 {
 public:
 	Dragons();
-	Dragons(b2World & world, int width, sf::Vector2f postion, bool circle) : Moveable(world, width, postion, circle) {}
+	Dragons(b2World & world, int width, sf::Vector2i postion, bool circle) : Moveable(world, width, postion, circle) {}
 	~Dragons();
 	virtual void move();
 };
