+++ Project/Drogon.cpp
@@ -1,6 +1,6 @@
 #include "Drogon.h"
 
-Drogon::Drogon(b2World & world, int width, sf::Vector2f postion, bool circle)
+Drogon::Drogon(b2World & world, int width, sf::Vector2i postion, bool circle)
 	: Dragons(world, width, postion, circle)
 {
 	m_sprite.setTexture(*Graphics::getInstance().getTexture(4));
