+++ Project/Guards.cpp
@@ -6,7 +6,7 @@ Guards::Guards()
 {
 }
 
-Guards::Guards(b2World & world, int width, sf::Vector2f postion, bool circle)
+Guards::Guards(b2World & world, int width, sf::Vector2i postion, bool circle)
 	: NonMoveable(world, width, postion, circle)
 {
 	m_sprite.setTexture(*Graphics::getInstance().getTexture(3));
